+++ ride-service/src/main/kotlin/io/voitovich/yura/rideservice/service/RideDriverManagementService.kt
@@ -1,0 +1,17 @@
+package io.voitovich.yura.rideservice.service
+
+import io.voitovich.yura.rideservice.dto.request.AcceptRideRequest
+import io.voitovich.yura.rideservice.dto.request.GetAvailableRidesRequest
+import io.voitovich.yura.rideservice.dto.request.UpdatePositionRequest
+import io.voitovich.yura.rideservice.dto.responce.GetAvailableRidesResponse
+import io.voitovich.yura.rideservice.dto.responce.RideResponse
+import io.voitovich.yura.rideservice.dto.responce.UpdatePositionResponse
+
+interface RideDriverManagementService {
+
+    fun updateDriverPosition(updatePositionRequest: UpdatePositionRequest): UpdatePositionResponse
+
+    fun getAvailableRides(getAvailableRidesRequest: GetAvailableRidesRequest) : GetAvailableRidesResponse
+
+    fun acceptRide(acceptRideRequest: AcceptRideRequest): RideResponse
+}
