+++ ride-service/src/main/kotlin/io/voitovich/yura/rideservice/service/RideService.kt
@@ -9,15 +9,4 @@ interface RideService {
     fun deleteRideById(id: UUID)
     fun getRidePage(pageRideRequest: RidePageRequest): RidePageResponse
 
-    fun createRide(request: CreateRideRequest) : CreateRideResponse
-
-    fun getAvailableRides(getAvailableRidesRequest: GetAvailableRidesRequest) : GetAvailableRidesResponse
-
-    fun acceptRide(acceptRideRequest: AcceptRideRequest): RideResponse
-
-    fun updateDriverPosition(updatePositionRequest: UpdatePositionRequest): UpdatePositionResponse
-
-    fun updatePassengerPosition(updatePositionRequest: UpdatePositionRequest): UpdatePositionResponse
-
-    fun cancelRide(cancelRequest: CancelRequest)
 }
